+++ mmdeploy/apis/utils/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .calibration import create_calib_table
+from .utils import build_task_processor, get_predefined_partition_cfg
+
+__all__ = [
+    'create_calib_table', 'build_task_processor',
+    'get_predefined_partition_cfg'
+]
